Week-3/backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
yfinance==0.2.66
requests==2.31.0
pandas
numpy
google-generativeai==0.3.2
python-dotenv==1.0.0
gunicorn==21.2.0